+++ src/Core/Methods/IosMethods.js
@@ -21,4 +21,31 @@ export default class IosMethods {
 
     this.nativeDeepWall.hidePaywallLoadingIndicator();
   }
+
+  /**
+   * requestAppTracking
+   *
+   * @param actionKey
+   * @param extraData
+   */
+  requestAppTracking(actionKey, extraData = null) {
+    if (!this.isIos()) {
+      return;
+    }
+
+    this.nativeDeepWall.requestAppTracking(actionKey, extraData);
+  }
+
+  /**
+   * sendExtraDataToPaywall
+   *
+   * @param extraData
+   */
+  sendExtraDataToPaywall(extraData) {
+    if (!this.isIos()) {
+      return;
+    }
+
+    this.nativeDeepWall.sendExtraDataToPaywall(extraData);
+  }
 }